+++ R/onLoad.R
@@ -1,0 +1,4 @@
+.onLoad <- function(libname, pkgname){
+    options(W_matreex_edist = TRUE)
+}
+